Implement Combat Phase
Revised movement so that new moves are created when the minimum stop speed is reached. Get basic attacking for mutants in and placeholder functions for the party to damage mutants - need to figure out how to delay each attack call.

Generalize queries in some components to avoid depending on column order from retrieving all columns. Some may be left behind due to convenience to access based on an index and loop